--- retrofit2/src/main/java/com/linecorp/armeria/client/retrofit2/metric/package-info.java
@@ -1,23 +1,0 @@
-/*
- * Copyright 2019 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-/**
- * Metric utilities for <a href="https://square.github.io/retrofit/">Retrofit2</a> adapter for armeria.
- */
-@NonNullByDefault
-package com.linecorp.armeria.client.retrofit2.metric;
-
-import com.linecorp.armeria.common.util.NonNullByDefault;
